Intern symbols in a simple hash table
This patch stores symbols in a simple hash table instead of a list,
speeding up symbol->string.

The structure is complicated enough though that it makes sense for the
garbage collector to add the symbol table to the root set.  Given that
we have no static data and the heap image doesn't leak from one
bootstrap to the next, this seems to be acceptable.

The improvement on bootstrap times is large:

```
Compiling stage1 compiler...
  Time: 73.408 s

Compiling stage2 compiler...
  Time: 71.36 s

Compiling stage3 compiler...
  Time: 5.733 s

Percent change (1 -> 3): -92%
Percent change (2 -> 3): -92%
```