Allow access to top-levels and intrinsics by value
This will allow e.g. (map cdr ls).  The strategy is to eta-expand all
top-level references, and then simplify those that are applied directly.

This will also eventually allow for arity checks in only those cases
where it's necessary (indirect calls), by reifying indirect calls for
calls to known callees but with the wrong number of arguments.

We remove i32.ne support, as it's not needed.  Also the i32.not
compilation was busted; there is no such wasm opcode.

This change will also allow for the compiler to automatically insert
adapters for external calls, for example to convert from wasm numbers to
JS numbers, and vice versa.  (We'd do this by putting the conversions in
the eta-expanded intrinsics.)

Compile time surprisingly does well:

```
Compiling stage1 compiler...
  Time: 6.198 s

Compiling stage2 compiler...
  Time: 5.47 s

Compiling stage3 compiler...
  Time: 4.166 s

Percent change (1 -> 3): -33%
Percent change (2 -> 3): -24%
```

I think this is because of putting intrinsics on the environment, so we
don't have to cons them up if a top-level isn't bound.