Rework to use JavaScript heap instead of linear memory
This commit changes Schism to allocate its objects on an external heap,
via call-outs to run-time imports.

Currently, references from wasm to external values is by "handle" -- an
index into a table.  In the future this will be refactored to use the
"anyref" facility.  In some ideal future, instead we use GC types, so
that we can avoid callouts to the runtime.