Explicitly represent zero-valued continuations via "drop"
By default in the dialect of Scheme implemented by Schism, all
expressions evaluate to a value.  Values that are in the head of a seq
are implicitly dropped.  This PR changes to make this drop explicit, and
as a corrolary to make the `head` of a `seq` node be a nullary
continuation.

Explicitly representing the drop allows the compiler to be a bit smarter
about how it deals with adapting between zero-valued and one-valued
continuations.  In the future it will allow for efficient invocation of
void-returning wasm imports.

In this patch, sadly, we have to distinguish between different types of
"if".  This patch adds if/void for "if" nodes whose arms are nullary.
Doing this allows us to simplify the compiler a bit.