Allow for typed imports
This commit adds support for typed imports.  The only difference
currently is functions that return void or not, and functions that
return boolean values or not.  For these functions, we add adapters so
that their eta-expanded values can deal in SCM values, inserting
synthetic (void) or #t/#f expressions as appropriate.

In the future we'll be able to call out to the runtime with untagged
numbers, receive untagged numbers, and eventually "anyref".

This commit does remove %wasm-import from the user vocabulary, for
better or worse; we can add it back in later if we want.  Feedback
welcome.