Use WebAssembly reference types to represent Scheme values
This commit represents Scheme types as native JavaScript values, without
indirecting through a table of "handles".  While we're at it, we'll
actually use straight-up JavaScript numbers and strings, although tagged
to partition between integers and characters on the one hand, and
strings and symbols on the other.  We'll also use normal JS values for
booleans and so forth.

This entirely removes the need for Schism to have a GC; the JS GC
handles it all, including the symbol table (via string uniquification).