Add initial support for libraries
This commit lays some of the groundwork for loading libraries. It adds
a pass ahead of the parser to handle import clauses and load the
transitive closure of all libraries referenced. Then the parser is
updated to construct an import environment for each library based on
the clauses it imports.

Currently, it only supports loading the library given in the test
case. For all other imports, and empty library is generated.

Also, clauses such as `(only ...)` are not yet handled correctly.

Issue #19