Add filesystem wrapper for Schism.Engine
This wrapper is very simple now, with only enough to support Schism
loading imported libraries from files. As we add features such as file
ports, this interface will probably expand and be refactored.

We provide a simple implementation of the file system interface for
Node, as well as a null interface to at least allow Schism to load and
run on other environments.