Factor primitives into external libraries
This splits the big list of primitive functions into their proper
libraries. The majority of them go into `(rnrs)`. The pair mutator
functions go into `(rnrs mutable-pairs)`. We had a handful of
non-standard functions that go into `(schism)`, which is somewhat
analogous to the `(chezscheme)` library that Schism used to pull a few
functions from.

Issue #19